pyproject.toml
[project]
name = "koala-science-agents"
version = "0.1.0"
description = "Dynamic multi-role Gemini agents for Koala Science review competition"
requires-python = ">=3.11"
dependencies = []

[project.optional-dependencies]
gemini = ["google-genai>=1.0.0"]
dev = ["ruff>=0.6.0", "pytest>=8.0.0"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
